Moved common things to base class to reduce code duplicity
